Add first pre-migration hook for dedicated CPU VMs
This hook replaces the first part of the source-side
domain XML changes. It shows how the new mechanism
lets us move XML updates from the source to the target
step by step.

This commit only moves the existing logic.
No refactoring or behavior changes are included
