+++ pkg/virt-launcher/premigration-hook-server/compute/BUILD.bazel
@@ -1,9 +1,36 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
     srcs = ["compute.go"],
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/premigration-hook-server/compute",
     visibility = ["//visibility:public"],
-    deps = ["//pkg/virt-launcher/premigration-hook-server/types:go_default_library"],
+    deps = [
+        "//pkg/handler-launcher-com/cmd/v1:go_default_library",
+        "//pkg/virt-launcher/premigration-hook-server/types:go_default_library",
+        "//pkg/virt-launcher/virtwrap/api:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter/vcpu:go_default_library",
+        "//pkg/virt-launcher/virtwrap/libvirtxml:go_default_library",
+        "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+        "//staging/src/kubevirt.io/client-go/log:go_default_library",
+        "//vendor/libvirt.org/go/libvirtxml:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = [
+        "compute_suite_test.go",
+        "compute_test.go",
+    ],
+    race = "on",
+    deps = [
+        ":go_default_library",
+        "//pkg/handler-launcher-com/cmd/v1:go_default_library",
+        "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+        "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
+        "//vendor/github.com/onsi/gomega:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//vendor/libvirt.org/go/libvirtxml:go_default_library",
+    ],
 )
