+++ pkg/virt-launcher/virtwrap/BUILD.bazel
@@ -62,7 +62,6 @@ go_library(
         "//pkg/virt-launcher/virtwrap/device/hostdevice/sriov:go_default_library",
         "//pkg/virt-launcher/virtwrap/efi:go_default_library",
         "//pkg/virt-launcher/virtwrap/errors:go_default_library",
-        "//pkg/virt-launcher/virtwrap/libvirtxml:go_default_library",
         "//pkg/virt-launcher/virtwrap/network:go_default_library",
         "//pkg/virt-launcher/virtwrap/stats:go_default_library",
         "//pkg/virt-launcher/virtwrap/statsconv:go_default_library",
